went apeshit and changed a ton of stuff too much for 1 commit
Everything looks ugly. coded for 10 hours straight though the night so bear with me...
Added controls and removed the defaults.
added the 360 helper and implemented the functions needed for the keybinds most of em are placeholders.
changed circlecoords from 2d to 3d.

No idea why this works after tonight but it does...